GlobalExceptionHandler: methods from GlobalControllerAdvice ExceptionController GlobalExceptionHandlerTest
