+++ src/main/java/school/faang/user_service/exception/GlobalExceptionHandler.java
@@ -3,46 +3,193 @@ package school.faang.user_service.exception;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
+import java.util.function.Consumer;
 
-@RestControllerAdvice
 @Slf4j
+@RestControllerAdvice
 public class GlobalExceptionHandler {
 
+    private static final String INFO = "Info";
+
     @ExceptionHandler(DataValidationException.class)
-    public ResponseEntity<String> handleDataValidationException(DataValidationException e) {
-        log.error("Data Validation exception occurred", e);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Map<String, String> handleDataValidationException(DataValidationException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put("Info", "Data Validation exception occurred");
+
+        logging(LoggingLevel.WARN, errors, e);
+
+        return errors;
     }
 
     @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<String> handleEntityNotFoundException(EntityNotFoundException e) {
-        log.error("Entity not Found exception occurred\n", e);
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public Map<String, String> handleEntityNotFoundException(EntityNotFoundException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Entity not Found exception occurred");
+
+        logging(LoggingLevel.WARN, errors);
+        return errors;
     }
 
     @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        log.error("Illegal argument exception occurred\n", e);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Map<String, String> handleIllegalArgumentException(IllegalArgumentException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Illegal argument exception occurred");
+
+        logging(LoggingLevel.WARN, errors);
+        return errors;
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, String>> handleValidationExceptions(MethodArgumentNotValidException e) {
-        Map<String, String> errors = e.getBindingResult().getFieldErrors().stream()
-                .collect(Collectors.toMap(
-                        FieldError::getField,
-                        error -> Objects.requireNonNullElse(error.getDefaultMessage(), "Unknown error"),
-                        (existing, replacement) -> existing
-                ));
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errors);
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Map<String, String> argumentNotValidExceptionHandler(MethodArgumentNotValidException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Data Validation exception occurred");
+
+        e.getBindingResult()
+                .getFieldErrors()
+                .forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
+
+        logging(LoggingLevel.WARN, errors);
+        return errors;
+    }
+
+    @ExceptionHandler(jakarta.validation.ConstraintViolationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Map<String, String> constraintViolationExceptionHandler(jakarta.validation.ConstraintViolationException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Data Validation exception occurred");
+
+        e.getConstraintViolations()
+                .forEach(violation -> {
+                    errors.put(violation.getRootBeanClass().toString(), violation.getRootBean().toString());
+                    errors.put(violation.getPropertyPath().toString(), violation.getMessage());
+                });
+
+        logging(LoggingLevel.WARN, errors);
+        return errors;
+    }
+
+    @ExceptionHandler(org.springframework.dao.DataIntegrityViolationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Map<String, String> constraintViolationExceptionHandler(org.springframework.dao.DataIntegrityViolationException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Data Validation exception occurred");
+
+        Throwable rootCause = e.getRootCause();
+        if (null == rootCause) {
+            return errors;
+        }
+        errors.put("Root cause", rootCause.toString());
+
+        if (rootCause instanceof org.hibernate.exception.ConstraintViolationException hibernateEx) {
+            String sqlMessage = hibernateEx.getSQLException().getMessage();
+            String constraintName = hibernateEx.getConstraintName();
+            errors.put("SQL error", sqlMessage);
+            errors.put("Violation of validation", constraintName);
+        } else if (rootCause instanceof java.sql.SQLException sqlEx) {
+            errors.put("SQL error", sqlEx.getMessage());
+        } else {
+            errors.put("Unknown error", e.getMessage());
+        }
+
+        logging(LoggingLevel.WARN, errors);
+        return errors;
+    }
+
+    @ExceptionHandler(org.hibernate.JDBCException.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public Map<String, String> handleJDBCException(org.hibernate.JDBCException e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "JDBC exception occurred");
+
+        java.sql.SQLException sqlException = e.getSQLException();
+        String sqlMessage = sqlException.getMessage();
+        int errorCode = sqlException.getErrorCode();
+        String sqlState = sqlException.getSQLState();
+
+        String title = "SQLException ";
+        errors.put(title + "error", "Database error occurred");
+        errors.put(title + "message", sqlMessage);
+        errors.put(title + "code", String.valueOf(errorCode));
+        errors.put(title + "state", sqlState);
+
+        logging(LoggingLevel.ERROR, errors, e);
+        return errors;
+    }
+
+    @ExceptionHandler(Exception.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public Map<String, String> otherExceptionHandler(Exception e) {
+        Map<String, String> errors = getErrorsMapWithExceptionTitle(e);
+
+        errors.put(INFO, "Exception occurred");
+
+        String cause = "Unknown";
+        if (null != e.getCause()) {
+            cause = e.getCause().toString();
+        }
+
+        errors.put("Cause", cause);
+
+        logging(LoggingLevel.ERROR, errors, e);
+        return errors;
+    }
+
+
+    private Map<String, String> getErrorsMapWithExceptionTitle(Exception e) {
+        Map<String, String> errors = new LinkedHashMap<>();
+
+        errors.put("Exception", e.getClass().toString());
+        errors.put("Exception message", e.getMessage());
+
+        return errors;
+    }
+
+
+    private void logging(LoggingLevel level, Map<String, String> errors) {
+        errors.entrySet().forEach(entry -> logging(level, entry, null));
+    }
+
+    private void logging(LoggingLevel level, Map<String, String> errors, Exception e) {
+        errors.entrySet().forEach(entry -> logging(level, entry, e));
+    }
+
+    private void logging(LoggingLevel level, Map.Entry<String, String> error, Exception e) {
+        String textForLogging = String.format("%s: %s", error.getKey(), error.getValue());
+        if (null != e) {
+            textForLogging = String.format("%s%n%s", textForLogging, e);
+        }
+        logging(level, textForLogging);
+    }
+
+    private void logging(LoggingLevel level, String errorMessage) {
+        level.command.accept(errorMessage);
+    }
+
+
+    public enum LoggingLevel {
+        INFO(log::info), DEBUG(log::debug), WARN(log::warn), ERROR(log::error);
+
+        final Consumer<String> command;
+        LoggingLevel(Consumer<String> command) {
+            this.command = command;
+        }
     }
 }
