+++ src/test/java/school/faang/user_service/exception/GlobalExceptionHandlerTest.java
@@ -1,0 +1,7 @@
+package school.faang.user_service.exception;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class GlobalExceptionHandlerTest {
+
+}
